requirements.txt
# Core Web Frameworks
FastAPI==0.109.0
uvicorn[standard]==0.27.0
Flask==3.0.0
Werkzeug==3.0.1

# API & Data Processing
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2

# ML & Embeddings
sentence-transformers>=3.0.0
faiss-cpu==1.7.4
torch>=2.3.0
scikit-learn==1.3.2
huggingface-hub>=0.19.3
transformers>=4.40.0,<5

# AI & LLM
google-generativeai==0.3.0

# Database
mysql-connector-python==8.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Optional
streamlit==1.30.0
